backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
slowapi
pydantic-settings
python-multipart
